+++ stag_lib/solve.cpp
@@ -15,8 +15,12 @@ double solution_error(const SprsMat* A, DenseVec& b, DenseVec& x) {
 }
 
 DenseVec stag::solve_laplacian(Graph* g, DenseVec& b, double eps) {
-  return solve_laplacian_jacobi(g, b, eps);
+  return solve_laplacian_gauss_seidel(g, b, eps);
 }
+
+//------------------------------------------------------------------------------
+// Jacobi iteration
+//------------------------------------------------------------------------------
 
 DenseVec stag::solve_laplacian_jacobi(stag::Graph* g, DenseVec& b, double eps,
                                       stag_int max_iterations) {
@@ -77,6 +81,60 @@ DenseVec stag::jacobi_iteration(const SprsMat* A, DenseVec& b, double eps,
   stag_int iteration = 0;
   while (error > eps) {
     x_t = x_t + P_inv * (b - *A * x_t);
+
+    error = solution_error(A, b, x_t);
+
+    // Check whether we have passed the maximum number of iterations.
+    iteration++;
+    if (iteration > max_iterations) throw stag::ConvergenceError();
+  }
+
+  return x_t;
+}
+
+//------------------------------------------------------------------------------
+// Gauss-Seidel Method
+//------------------------------------------------------------------------------
+
+DenseVec stag::solve_laplacian_gauss_seidel(stag::Graph* g, DenseVec& b,
+                                            double eps,
+                                            stag_int max_iterations) {
+  DenseVec x = stag::gauss_seidel_iteration(g->laplacian(), b, eps,
+                                            max_iterations);
+  return x;
+}
+
+DenseVec stag::solve_laplacian_gauss_seidel(stag::Graph* g, DenseVec& b,
+                                            double eps) {
+  return stag::solve_laplacian_gauss_seidel(g, b, eps, STAG_MAX_ITERATIONS);
+}
+
+DenseVec stag::gauss_seidel_iteration(const SprsMat *A,
+                                      DenseVec &b,
+                                      double eps,
+                                      stag_int max_iterations) {
+  // We will make use of the upper triangular part of A
+  Eigen::TriangularView<const SprsMat, Eigen::StrictlyUpper> U =
+      A->triangularView<Eigen::StrictlyUpper>();
+
+  // In each iteration, we compute the solution to
+  //    L x_{k+1} = b - U x_k,
+  // where L is the lower triangular part of A.
+  // We will iterate until the error is below the provided epsilon.
+  DenseVec x_t = b;
+  double error = solution_error(A, b, x_t);
+  stag_int iteration = 0;
+  while (error > eps) {
+    // Solve the lower-triangular system
+    //   L x = b - U x
+    DenseVec bp = b - U * x_t;
+    for (auto i = 0; i < A->rows(); i++) {
+      double target = bp.coeff(i);
+      for (auto j = 0; j < i; j++) {
+        target -= A->coeff(i, j) * x_t.coeff(j);
+      }
+      x_t.coeffRef(i) = target / A->coeff(i,i);
+    }
 
     error = solution_error(A, b, x_t);